Refactor(soonki/club): Frame.js -> index.js 변경
