+++ src/components/Club/Home/Activities/Activities.js
