+++ src/components/Club/Home/Intro/Skeleton.js
