+++ src/components/Club/Home/Manager/index.js
