+++ src/components/Club/Home/index.js
@@ -3,16 +3,16 @@ import { getClubInfo } from 'redux/slices/clubhome';
 import { useRouter } from 'next/router';
 import { useDispatch, useSelector } from 'react-redux';
 import { changeComp } from 'redux/slices/chageComp';
-import ClubIntro from '../Intro/ClubIntro';
-import ClubNotice from '../Notice/ClubNotice';
-import Activities from '../Activities/Activities';
-import Calendar from '../Calendar';
-import Review from '../Review/Review';
-import Apply from '../Apply';
+import ClubIntro from './Intro/ClubIntro';
+import ClubNotice from './Notice/ClubNotice';
+import Activities from './Activities/Activities';
+import Calendar from './Calendar';
+import Review from './Review/Review';
+import Apply from './Apply';
 import SideBar from './SideBar';
 import styles from 'styles/Club/Home/Common/frame.module.scss';
 
-const Frame = () => {
+const Club = () => {
   const [isVisit, setIsVisit] = useState(false);
 
   const dispatch = useDispatch();
@@ -58,4 +58,4 @@ const Frame = () => {
   );
 };
 
-export default Frame;
+export default Club;
