+++ src/pages/clubhome/[id]/index.js
@@ -1,5 +1,5 @@
+import Club from 'components/Club/Home';
 import Head from 'next/head';
-import Frame from 'components/Club/ClubHome/Common/frame';
 
 const ClubHome = () => {
   return (
@@ -7,7 +7,7 @@ const ClubHome = () => {
       <Head>
         <title>동그라미 | 동아리 홈</title>
       </Head>
-      <Frame />
+      <Club />
     </>
   );
 };
